Update: remove required condition from form schema
