--- app/views/AssessmentRegistryForm/styles.css
@@ -1,6 +1,0 @@
-.tab-panel {
-    display: flex;
-    flex-direction: column;
-    flex-grow: 1;
-    padding: var(--dui-spacing-large);
-}
