+++ app/views/EditAry/AssessmentRegistryForm/MetadataForm/StakeholderForm/index.tsx
@@ -7,8 +7,9 @@ import AddStakeholderButton from '#components/general/AddStakeholderButton';
 import StakeholderList from '#views/ProjectEdit/ProjectDetailsForm/StakeholderList';
 import { ProjectOrganizationTypeEnum } from '#generated/types';
 import { BasicOrganization } from '#types';
-import { BasicProjectOrganization, PartialFormType } from '#views/AssessmentRegistryForm/formSchema';
 import { getErrorObject } from '#components/framework/AttributeInput';
+
+import { BasicProjectOrganization, PartialFormType } from '../../../AssessmentRegistryForm/formSchema';
 
 import styles from './styles.css';
