+++ app/views/EditAry/AssessmentRegistryForm/formSchema.ts
@@ -34,21 +34,18 @@ export const schema: FormSchema = {
         frequency: [requiredCondition],
         confidentiality: [requiredCondition],
         language: [requiredCondition],
-        noOfPages: [requiredCondition],
+        noOfPages: [],
         dataCollectionStartDate: [],
         dataCollectionEndDate: [],
-        publicationDate: [requiredCondition],
-        leadOrganizations: [requiredCondition],
-        internationalPartners: [requiredCondition],
-        donors: [requiredCondition],
-        nationalPartners: [requiredCondition],
-        governments: [requiredCondition],
+        publicationDate: [],
+        leadOrganizations: [defaultEmptyArrayType],
+        internationalPartners: [defaultEmptyArrayType],
+        donors: [defaultEmptyArrayType],
+        nationalPartners: [defaultEmptyArrayType],
+        governments: [defaultEmptyArrayType],
 
         // NOTE: uncomment on other form
         locations: [defaultEmptyArrayType],
-        // focuses: [defaultEmptyArrayType],
-        // sectors: [defaultEmptyArrayType],
-        // protectionInfoMgmts: [defaultEmptyArrayType],
         methodologyAttributes: [defaultEmptyArrayType],
         additionalDocuments: [defaultEmptyArrayType],
         scoreRatings: [defaultEmptyArrayType],
