+++ app/views/EditAry/AssessmentRegistryForm/styles.css
@@ -1,0 +1,22 @@
+.assessment-registry-form {
+    display: flex;
+    flex-direction: column;
+
+    .tab-list {
+        display: flex;
+        flex-shrink: 0;
+
+        .dummy {
+            flex-grow: 1;
+            border-bottom: var(--dui-width-separator-medium) solid var(--dui-color-separator);
+        }
+    }
+
+    .tab-panel {
+        display: flex;
+        flex-direction: column;
+        flex-grow: 1;
+        padding: var(--dui-spacing-large);
+
+    }
+}
