+++ app/views/EditAry/index.tsx
@@ -19,10 +19,8 @@ import BackLink from '#components/BackLink';
 import SubNavbar from '#components/SubNavbar';
 import ProjectContext from '#base/context/ProjectContext';
 import LeftPaneEntries from '#components/LeftPaneEntries';
-import AssessmentRegistryForm from '#views/AssessmentRegistryForm';
 import { Entry, EntryInput as EntryInputType } from '#components/entry/types';
 import { ORGANIZATION_FRAGMENT, ENTRY_FRAGMENT } from '#gqlFragments';
-import { initialValue, schema } from '#views/AssessmentRegistryForm/formSchema';
 import {
     AssessmentRegistryCreateInputType,
     CreateAssessmentRegistryMutation,
@@ -32,6 +30,9 @@ import {
     EntriesFromAssessmentQuery,
     EntriesFromAssessmentQueryVariables,
 } from '#generated/types';
+
+import AssessmentRegistryForm from './AssessmentRegistryForm';
+import { initialValue, schema } from './AssessmentRegistryForm/formSchema';
 
 import styles from './styles.css';
